Split out exit function and codes into a separate module
With _error_reporting getting more complicated, the obvious thing to move
to a separate module is the exit functionality.
